remove dependency on serialize.h and util.h for SecureString
