+++ src/crypter.h
@@ -4,7 +4,7 @@
 #ifndef __CRYPTER_H__
 #define __CRYPTER_H__
 
-#include "util.h" /* for SecureString */
+#include "allocators.h" /* for SecureString */
 #include "key.h"
 
 const unsigned int WALLET_CRYPTO_KEY_SIZE = 32;
